+++ packages/rspack/tests/diagnostics/module-build-failed/loader-throw-error/stats.err
@@ -1,1 +1,1 @@
-ERROR in ./lib.js× Module build failed:╰─▶   × Error: Failed to loadat Object.<anonymous>.module.exports (<PROJECT_ROOT>/tests/diagnostics/module-build-failed/loader-throw-error/my-loader.js:2:9)at apply (<PROJECT_ROOT>/src/loader-runner/index.ts:741:14)at runSyncOrAsync (<PROJECT_ROOT>/src/loader-runner/index.ts:742:3)at runSyncOrAsync (<PROJECT_ROOT>/src/loader-runner/index.ts:864:3)at callback (<PROJECT_ROOT>/src/loader-runner/loadLoader.js:152:2)at handleResult (<PROJECT_ROOT>/src/loader-runner/loadLoader.js:122:10)at loadLoader (<PROJECT_ROOT>/src/loader-runner/index.ts:850:2)at iterateNormalLoaders (<PROJECT_ROOT>/src/loader-runner/index.ts:655:4)at new Promise (<anonymous>)at runLoaders (<PROJECT_ROOT>/src/loader-runner/index.ts:630:9)help: File was processed with this loader: '<PROJECT_ROOT>/tests/diagnostics/module-build-failed/loader-throw-error/my-loader.js'
+ERROR in ./lib.js× Module build failed:╰─▶   × Error: Failed to loadat Object.<anonymous>.module.exports (<PROJECT_ROOT>/tests/diagnostics/module-build-failed/loader-throw-error/my-loader.js:2:9)at apply (<PROJECT_ROOT>/dist/loader-runner/index.js:616:23)at runSyncOrAsync (<PROJECT_ROOT>/dist/loader-runner/index.js:617:9)at runSyncOrAsync (<PROJECT_ROOT>/dist/loader-runner/index.js:721:9)at callback (<PROJECT_ROOT>/dist/loader-runner/loadLoader.js:128:5)at handleResult (<PROJECT_ROOT>/dist/loader-runner/loadLoader.js:109:16)at loadLoader (<PROJECT_ROOT>/dist/loader-runner/index.js:710:5)at iterateNormalLoaders (<PROJECT_ROOT>/dist/loader-runner/index.js:537:13)at new Promise (<anonymous>)at runLoaders (<PROJECT_ROOT>/dist/loader-runner/index.js:512:12)help: File was processed with this loader: '<PROJECT_ROOT>/tests/diagnostics/module-build-failed/loader-throw-error/my-loader.js'
